requirements.txt
flake8==3.8.4
Flask==1.1.2
Jinja2==2.11.2
jsonschema==3.2.0
mypy==0.790
mypy-extensions==0.4.3
pep257==0.7.0
pycparser==2.20
python-dateutil==2.8.1
radon==4.3.2
SQLAlchemy==1.3.22
vulture==2.3

